cache: Support just the new glibc format
Some systems apparently actually have a ld.so.cache that is the new
format (and not the backwards compatible old+new format), so attempt
to handle the case.

Untested because, I do not have such a system, and messing up my
laptop's dynamic linker cache is not my idea of a fun time.